# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Auth Public OpenAPI",
    "version": "1.0.0",
    "description": "Confidential server-side applications only. HTTPS required. Application-level IP allowlist, explicit scopes, S256 PKCE and independent rotatable credentials. See openapi.md for proxy deployment and rollout requirements."
  },
  "servers": [
    {
      "url": "https://auth.example.com",
      "description": "Replace with the deployment HTTPS origin"
    }
  ],
  "paths": {
...
    "/openapi/v1/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "Read minimal current user profile",
        "security": [
          {
            "oauth": [
              "profile:read"
            ]
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid parameters, scope or grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid client or access token; WWW-Authenticate is included",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "HTTPS, application IP or scope policy refused access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Shared rate limit exceeded; retry after 60 seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "const": 60
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable; access refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "description": "Bearer token and application egress IP allowlist required. No query parameters or arbitrary user/app selectors are accepted. All responses are no-store."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "clientBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "APP_ID:OPENAPI_SECRET (oas_...), never the legacy RPC secret"
      },
      "browserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "u",
        "description": "Existing Auth browser session, only at authorization endpoint"
      },
      "oauth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/openapi/v1/oauth/authorize",
            "tokenUrl": "/openapi/v1/oauth/token",
            "scopes": {
              "profile:read": "Minimal current user profile",
              "permissions:read": "Current application roles and permissions",
              "permissions:check": "Check a current application permission"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "additionalProperties": false
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "avatar"
        ],
        "additionalProperties": false
      },
...
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Profile"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "additionalProperties": false
      },
...
    }
  }
}